Rename ForkManager to ForkControl #98
Actually it should be process control, but it just
abstracts the forking part.
