+++ src/Common/ForkControl/Exception/ForkControlException.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Gaming\Common\ForkControl\Exception;
+
+use Exception;
+
+final class ForkControlException extends Exception
+{
+}
