--- src/Common/ForkManager/Exception/ForkManagerException.php
@@ -1,11 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Gaming\Common\ForkManager\Exception;
-
-use Exception;
-
-final class ForkManagerException extends Exception
-{
-}
