+++ src/Common/Port/Adapter/Symfony/Worker.php
@@ -5,8 +5,8 @@ declare(strict_types=1);
 namespace Gaming\Common\Port\Adapter\Symfony;
 
 use Gaming\Common\EventStore\StoredEventSubscriber;
-use Gaming\Common\ForkManager\Process;
-use Gaming\Common\ForkManager\Task;
+use Gaming\Common\ForkControl\Process;
+use Gaming\Common\ForkControl\Task;
 
 final class Worker implements Task
 {
